rag-demo/pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[vertexai]>=1.26.0",
    "google-cloud-aiplatform>=1.140.0",
    "python-dotenv>=1.2.2",
]
